+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.YY.Z (2023-MM-DD)
 
 - CI: Don't run tests when only `__init__` or `__meta__` is updated
+- CI: Move data elsewhere than git
 
 ## 1.23.0 (2023-01-09)